Refactor album data management into a custom hook using the singleton pattern
