+++ src/containers/albumContainer/albumContainer.js
@@ -4,50 +4,31 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
 import spotifyLogo from "../../assets/spotify_Primary_Logo_White_RGB.svg";
 import { useLocation } from "react-router-dom";
-import { getCachedEntry } from "../../utilities/indexedDb";
+import { useAlbumData } from "../../hooks/useAlbumData";
 import { logger } from "../../utilities/logger";
 import { useNavigationHelpers } from "../../utilities/navigationHelpers";
 
 function AlbumContainer() {
     const location = useLocation();
-    const [album, setAlbum] = useState(null);
-    const [loading, setLoading] = useState(true);
+    const { groupedAlbums } = useAlbumData();
     const { goTo } = useNavigationHelpers();
 
-    // Extract albumId and genre from query params
+    // Extract params
     const params = new URLSearchParams(location.search);
     const albumId = params.get("albumId");
     const genre = params.get("genre");
     const albumSearch = params.get("albumSearch") || "";
     const genreSearch = params.get("genreSearch") || "";
 
-    useEffect(() => {
-        async function fetchAlbum() {
-            setLoading(true);
-            try {
-                // Try to find the album in cached grouped_albums
-                const groupedAlbums = await getCachedEntry('data', 'grouped_albums');
-                let found = null;
-                let foundGenre = null;
-                if (groupedAlbums) {
-                    for (const [genre, albums] of Object.entries(groupedAlbums)) {
-                        found = albums.find(a => a.id === albumId);
-                        if (found) {
-                            foundGenre = genre;
-                            break;
-                        }
-                    }
-                }
-                setAlbum(found ? { ...found, _genre: foundGenre } : null);
-            } catch (e) {
-                logger.error('ALBUM001', 'Error fetching album', { albumId, error: e });
-            }
-            setLoading(false);
-        }
-        if (albumId) fetchAlbum();
-    }, [albumId]);
+    // Find album directly in the specified genre
+    const album = React.useMemo(() => {
+        if (!groupedAlbums || !albumId || !genre) return null;
+        const albums = groupedAlbums[genre];
+        if (!albums) return null;
+        const found = albums.find(a => a.id === albumId);
+        return found ? { ...found, _genre: genre } : null;
+    }, [groupedAlbums, albumId, genre]);
 
-    if (loading) return <div className="single-album-container">Loading...</div>;
     if (!album) return <div className="single-album-container">Album not found.</div>;
 
     return (
