+++ src/containers/genreContainer/genreContainer.js
@@ -1,44 +1,26 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import SearchSortContainer from "../../components/SearchSortContainer";
 import "./genreContainer.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
 import { useNavigationHelpers } from "../../utilities/navigationHelpers";
 import { useLocation } from "react-router-dom";
-import { getCachedEntry } from "../../utilities/indexedDb";
-import { logger } from "../../utilities/logger";
+import { useAlbumData } from "../../hooks/useAlbumData";
 
 function GenreContainer() {
     const location = useLocation();
     const { goTo } = useNavigationHelpers();
+    const { groupedAlbums } = useAlbumData();
 
     const params = new URLSearchParams(location.search);
     const genre = params.get("genre");
     const genreSearch = params.get("genreSearch") || "";
     const albumSearch = params.get("albumSearch") || "";
 
-    const [albums, setAlbums] = useState([]);
     const [searchQuery, setSearchQuery] = useState(albumSearch || "");
     const [sortOption, setSortOption] = useState("alphabetical-asc-artist");
-    const [loading, setLoading] = useState(true);
 
-    useEffect(() => {
-        async function fetchAlbums() {
-            setLoading(true);
-            try {
-                const groupedAlbums = await getCachedEntry('data', 'grouped_albums');
-                let genreAlbums = [];
-                if (groupedAlbums && groupedAlbums[genre]) {
-                    genreAlbums = groupedAlbums[genre];
-                }
-                setAlbums(genreAlbums);
-            } catch (e) {
-                logger.error('GENRE001', 'Error fetching genre albums', { genre, error: e });
-            }
-            setLoading(false);
-        }
-        if (genre) fetchAlbums();
-    }, [genre]);
+    const albums = groupedAlbums?.[genre] || [];
 
     const sortOptions = [
         { value: "alphabetical-asc-album", label: "A-Z (Album)" },
@@ -68,7 +50,6 @@ function GenreContainer() {
         return 0;
     });
 
-    if (loading) return <div className="genre-container">Loading...</div>;
     if (!genre || !albums.length) return <div className="genre-container">No albums found for this genre.</div>;
 
     return (
